Move management tests (#7318)
